feat : 각 컴포넌트 주석 및 파일 이동
