+++ src/components/shared/ListItem.tsx
@@ -1,10 +1,36 @@
 import { listItem, type ListItemVariants } from "@/components/style/listItem";
 
+/**
+ *
+ * ListItem 컴포넌트
+ * 제목, 작성자, 설명, 좋아요 수, 댓글 수 등 리스트 형태로 콘텐츠를 보여주는 UI
+ * 의도(intent) 및 사이즈(size)에 따라 스타일이 달라지며, truncate 및 wrapping 적용
+ *
+ * @function ListItem.tsx
+ * @date 2025/05/31
+ * @history
+ * -------------------------------------------------------
+ *           변경일             작성자             변경내용
+ * -------------------------------------------------------
+ *
+ *        2025/05/31           이철욱               신규작성
+ *
+ * @param title 항목 제목
+ * @param writer 작성자
+ * @param description 항목 설명
+ * @param size 스타일 사이즈
+ * @param intent 스타일 의도
+ * @param className 추가 클래스
+ * @param likes 좋아요 수
+ * @param comment 댓글 수
+ */
+
 interface ListItemProps extends ListItemVariants {
   title: string;
   writer: string;
   description: string;
   className?: string;
+  likes: number;
   comment: number;
 }
 
@@ -15,6 +41,7 @@ const ListItem = ({
   size,
   intent,
   className,
+  likes,
   comment,
 }: ListItemProps) => {
   const combinedClass = `${listItem({ size, intent })} ${className ?? ""}`.trim();
@@ -25,13 +52,16 @@ const ListItem = ({
         <div className="font-black justify-start text-xl truncate">{title}</div>
       </div>
 
-      <div className="flex justify-between w-full">
+      <div className="flex-default w-full">
         <div className="flex flex-col justify-between">
           <p className="list-col-wrap text-base flex-1">{description}</p>
           <div className="text-xs uppercase font-semibold opacity-60">{writer}</div>
         </div>
         <div className="flex justify-end">
-          <button className="btn btn-ghost"></button>
+          <button className="btn btn-ghost">
+            {/* 여기에 댓글 icon, count 박을 예정 */}
+            <span>{comment}</span>
+          </button>
 
           <button className="btn btn-ghost">
             <svg className="size-[1.2em]" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">
@@ -46,7 +76,7 @@ const ListItem = ({
               </g>
             </svg>
 
-            <span>{comment}</span>
+            <span>{likes}</span>
           </button>
         </div>
       </div>
